Added vue.js and axios.js files to static content and updated index.html
The web service now delivers all files necessary to run the UI
locally, so that the service still works on a localhost display
without a network connection.

The REST service now supports saving and restoring the config files on
the thermostat boards.

Fixed a bug in handling of corrupt config files.